x location

+++ pdf/DataType.py
@@ -4,7 +4,7 @@ from dataclasses import dataclass
 class TextLine:
     text: str
     size: int            
-    location: float       # point where textline starts
+    location: int         # point where textline starts
     keyword_set: list     # based on changes of color or size
     page_num : int        # page number that textline came from
